EnemyFSM first layout
