+++ Assets/Scripts/Enemy/Behaviour/BehaviourConfig.cs
@@ -1,0 +1,10 @@
+using System;
+using UnityEngine;
+
+[Serializable]
+public class BehaviourConfig
+{
+    public BehaviourType behaviourType;
+    [SerializeReference] public IBehaviourData behaviourData;
+    public IBehaviour behaviour;
+}
