+++ Assets/Scripts/Enemy/Behaviour/BehaviourConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a23bdf922aa5cf14babdf4eaac7fb1c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
