+++ Assets/Scripts/Enemy/Behaviour/BehaviourData.cs
@@ -1,0 +1,13 @@
+using System;
+using UnityEngine;
+
+[Serializable]
+public class KnightBehaviourData : IBehaviourData
+{
+    //TODO : Set All the variables needed
+
+    [SerializeField] float data1 = 0;
+    [SerializeField] float data2 = 0;
+    [SerializeField] float data3 = 0;
+    [SerializeField] float data4 = 0;
+}
