+++ Assets/Scripts/Enemy/Behaviour/BehaviourData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69b82897d16bf7d4e9666f18cf906c60
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
