+++ Assets/Scripts/Enemy/Behaviour/BehaviourDataFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7d973d7b2cf52a42b09d25790d249c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
