+++ Assets/Scripts/Enemy/Behaviour/BehaviourFactory.cs
@@ -1,0 +1,12 @@
+/*public class BehaviourFactory
+{
+    public static IBehaviour CreateBehaviour(BehaviourType type)
+    {
+        return type switch
+        {
+            BehaviourType.Knight => ,
+            BehaviourType.Eye => new EnemyBehaviour(),
+            _ => null,
+        };
+    }
+}*/
