+++ Assets/Scripts/Enemy/Behaviour/IBehaviour.cs
@@ -1,0 +1,5 @@
+public interface IBehaviour
+{
+    public void Update() { }
+    public void InitBehaviour() { }
+}
