+++ Assets/Scripts/Enemy/Behaviour/IBehaviourData.cs
@@ -1,0 +1,4 @@
+public interface IBehaviourData
+{
+    
+}
