+++ Assets/Scripts/Enemy/Behaviour/IBehaviourData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dc891f50f2fd8e847bf4fb57b4a6fe8c
+guid: 06ac58226f9399547a1f1a6b25e045d4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
