+++ Assets/Scripts/Enemy/EnemyControllerOld.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac7d6a3285c40d3428af9a81b49f0ce6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
